msvcrt/tests: Fix signed char issues.

+++ dlls/msvcrt/tests/string.c
@@ -2810,13 +2810,13 @@ static void test_wctob(void)
     ok(ret == EOF, "ret = %x\n", ret);
 
     ret = p_wctob(0x81);
-    ok(ret == (int)(char)0x81, "ret = %x\n", ret);
+    ok(ret == (signed char)0x81, "ret = %x\n", ret);
 
     ret = p_wctob(0x9f);
-    ok(ret == (int)(char)0x9f, "ret = %x\n", ret);
+    ok(ret == (signed char)0x9f, "ret = %x\n", ret);
 
     ret = p_wctob(0xe0);
-    ok(ret == (int)(char)0xe0, "ret = %x\n", ret);
+    ok(ret == (signed char)0xe0, "ret = %x\n", ret);
 
     _setmbcp(cp);
 }
